pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unified-reasoning-rl"
version = "0.1.0"
description = "Unified Reinforcement Learning Framework for Mathematical Reasoning with PPO, GRPO, Dr.GRPO, GSPO, and DAPO"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Unified RL Team"}
]
keywords = [
    "reinforcement-learning",
    "llm",
    "mathematical-reasoning",
    "ppo",
    "grpo",
    "rlhf"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy<2.0.0",
    "torch>=2.4.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "bitsandbytes>=0.43.0",
    "accelerate>=0.30.0",
    "scipy",
    "pyyaml",
    "tqdm",
    "datasets>=2.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "black",
    "isort",
]
logging = [
    "wandb",
]

[project.urls]
Repository = "https://github.com/your-org/unified-reasoning-rl"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
